+++ Shell/Language.cs
@@ -22,6 +22,7 @@ namespace KillerShell.Shell
             (Locale.Fr,   "Français",    "fr-FR"),
             (Locale.Ja,   "日本語",       "ja-JP"),
             (Locale.PlPL, "Polski",       "pl-PL"),
+            (Locale.HuHU, "Magyar",       "hu-HU"),
             (Locale.TrTR, "Türkçe",      "tr-TR"),
             (Locale.ZhCN, "中文 (简体)",  "zh-CN"),
             (Locale.ZhTW, "中文 (繁體)",  "zh-TW"),